+++ server_test.go
@@ -1690,9 +1690,8 @@ func TestServer_DropContinuousQuery(t *testing.T) {
 	}
 }
 
-// Ensure
+// Ensure continuous queries run
 func TestServer_RunContinuousQueries(t *testing.T) {
-	t.Skip()
 	c := test.NewDefaultMessagingClient()
 	defer c.Close()
 	s := OpenServer(c)